+++ test/spec/CommonSpec.js
@@ -1,64 +1,83 @@
-
-describe('Common Utilities', function () {
+describe('Comparator Functions', function () {
     var hm = window.hm;
-    var types;
+    var types = {
+        'undefined': void 0,
+        'null': null,
+        'boolean': true,
+        'string': 'Hello World',
+        'number': 42,
+        'object': {'one': 1, 'two': 2, 'three': 3},
+        'function': function () {}
+    };
 
     beforeEach(function () {
-        types = {
-            'undefined': void 0,
-            'null': null,
-            'boolean': true,
-            'string': 'Hello World',
-            'number': 42,
-            'array': [1, 2, 3],
-            'object': {'one': 1, 'two': 2, 'three': 3},
-            'function': function () {}
-        };
+        var type;
+
+        jasmine.addMatchers({
+            toBeFalseForAllOtherTypes: function () {
+                return {
+                    compare: function (original) {
+                        var valid = true;
+                        var method, message;
+
+                        for (type in types) {
+                            if (types.hasOwnProperty(type)) {
+                                method = 'is' + (type.charAt(0).toUpperCase() + type.slice(1));
+                                if (method !== original && !hm[method](types[type])) {
+                                    valid = false;
+                                    message = "Expected '" + original + "' to be false for '" + type + "'";
+                                }
+                            }
+                        }
+
+                        return {
+                            pass: valid,
+                            message: message
+                        }
+                    }
+                }
+            }
+        });
     });
 
-    it('hm.isUndefined() should return true when variable is undefined', function () {
+    it('hm.isUndefined() returns true when argument is undefined', function () {
         expect(hm.isUndefined(types.undefined)).toBe(true);
-        expect('isUndefined').toBeFalseForAllOtherTypes(types);
+        expect('isUndefined').toBeFalseForAllOtherTypes();
     });
 
-    it('hm.isNull() should return true when variable is null', function () {
+    it('hm.isNull() should return true when argument is null', function () {
         expect(hm.isNull(types.null)).toBe(true);
-        expect('isNull').toBeFalseForAllOtherTypes(types);
+        expect('isNull').toBeFalseForAllOtherTypes();
     });
 
-    it('hm.isBoolean() should return true when variable is a boolean', function () {
+    it('hm.isBoolean() should return true when argument is a boolean', function () {
         expect(hm.isBoolean(types.boolean)).toBe(true);
-        expect('isBoolean').toBeFalseForAllOtherTypes(types);
+        expect('isBoolean').toBeFalseForAllOtherTypes();
     });
 
-    it('hm.isString() should return true when variable is a string', function () {
+    it('hm.isString() should return true when argument is a string', function () {
         expect(hm.isString(types.string)).toBe(true);
-        expect('isString').toBeFalseForAllOtherTypes(types);
+        expect('isString').toBeFalseForAllOtherTypes();
     });
 
-    it('hm.isNumber() should return true when variable is a number', function () {
+    it('hm.isNumber() should return true when argument is a number', function () {
         expect(hm.isNumber(types.number)).toBe(true);
-        expect('isNumber').toBeFalseForAllOtherTypes(types);
+        expect('isNumber').toBeFalseForAllOtherTypes();
     });
 
-    it('hm.isFunction() should return true when variable is a function', function () {
+    it('hm.isFunction() should return true when argument is a function', function () {
         expect(hm.isFunction(types.function)).toBe(true);
-        expect('isFunction').toBeFalseForAllOtherTypes(types);
+        expect('isFunction').toBeFalseForAllOtherTypes();
     });
 
-    it('hm.isArray() should return true when variable is an array', function () {
-        expect(hm.isArray(types.array)).toBe(true);
-        expect('isArray').toBeFalseForAllOtherTypes(types);
-    });
-
-    it('hm.isObject() should return true when variable is an object or array', function () {
-        expect(hm.isObject(types.array)).toBe(true);
+    it('hm.isObject() should return true when argument is an object or array', function () {
+        expect(hm.isObject([1, 2, 3])).toBe(true);
         expect(hm.isObject(types.object)).toBe(true);
+
     });
 
-    it('hm.isPlainObject() should return true when variable is a plain object', function () {
+    it('hm.isPlainObject() should return true when argument is a plain object', function () {
         expect(hm.isPlainObject(types.object)).toBe(true);
         expect(hm.isPlainObject(document.createElement('div'))).toBe(false);
     });
 });
-